sender: update eventexporter
